Fixed #30876 -- Moved classproperty() decorator to the django.utils.functional.
