+++ django/test/selenium.py
@@ -3,7 +3,7 @@ import unittest
 from contextlib import contextmanager
 
 from django.test import LiveServerTestCase, tag
-from django.utils.decorators import classproperty
+from django.utils.functional import classproperty
 from django.utils.module_loading import import_string
 from django.utils.text import capfirst
