+++ tests/utils_tests/test_decorators.py
@@ -2,7 +2,7 @@ from django.http import HttpResponse
 from django.template import engines
 from django.template.response import TemplateResponse
 from django.test import RequestFactory, SimpleTestCase
-from django.utils.decorators import classproperty, decorator_from_middleware
+from django.utils.decorators import decorator_from_middleware
 
 
 class ProcessViewMiddleware:
@@ -108,41 +108,3 @@ class DecoratorFromMiddlewareTests(SimpleTestCase):
         self.assertTrue(getattr(request, 'process_response_reached', False))
         # process_response saw the rendered content
         self.assertEqual(request.process_response_content, b"Hello world")
-
-
-class ClassPropertyTest(SimpleTestCase):
-    def test_getter(self):
-        class Foo:
-            foo_attr = 123
-
-            def __init__(self):
-                self.foo_attr = 456
-
-            @classproperty
-            def foo(cls):
-                return cls.foo_attr
-
-        class Bar:
-            bar = classproperty()
-
-            @bar.getter
-            def bar(cls):
-                return 123
-
-        self.assertEqual(Foo.foo, 123)
-        self.assertEqual(Foo().foo, 123)
-        self.assertEqual(Bar.bar, 123)
-        self.assertEqual(Bar().bar, 123)
-
-    def test_override_getter(self):
-        class Foo:
-            @classproperty
-            def foo(cls):
-                return 123
-
-            @foo.getter
-            def foo(cls):
-                return 456
-
-        self.assertEqual(Foo.foo, 456)
-        self.assertEqual(Foo().foo, 456)
